Projects Manager Early Alpha
Contains about 90% of the Projects Manager function for the projects
CPT. Everything work and is in place minus one feature and one known
bug. I began working on the step creation but pushed up the code before
I started that.

* Need to enable "Step Creator" - Allows us to create a new step where
we can add up to three images and a list of sub-steps.
* KNOWN BUG - When removing sub-lists, the add button doesn't work.
Found a patch, but introduced a new bug. Need to find the inital bug
and patch correctly.